requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
playwright==1.50.0
google-api-python-client==2.165.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pytz==2025.2
